Generate json file containing key build paths
The build system knows where all these components are, but the paths
used may vary. The name of the final qemu binary is dependent on the target
architecture. The path to the ATF directory depends on whether a debug or
release build was used.

The generated file is inteded to export this information for use when
running QEMU.

Bug: 122330825
Change-Id: I839dba2f3032a05d668894875bcc2b3eeaed8e1c