Enable pattern initialization in test builds
Clang recently added a new flag which will initialize variables with an
intentionally invalid (or likely to be invalid) pattern to catch
accidental assumptions about the initial contents of a variable.

This change enables that flag in test builds, and when combined with
UBSan should help detect these conditions should they arise.

Bug: 140579848
Change-Id: I28e4bae22399f2cf63d8c2c49cbe09492b333626